refactor: restructure engine, deduplicate OOB/Options/SDK logic
- Split engine.go (2045→878 lines) into 7 focused files:
  webprobe, pedm, stage_fingerprint, filter, oob_coordinator, context, errors
- Unify OOB token detection (ContainsOOBToken/PocUsesOOB exported)
- Extract IsNetOnly() method on poc.Poc
- Add RunnerOption pattern (WithSDKMode), delete createSDKRunner
- Add sendOrDrop[T] generic channel helper, integrated at 6 call sites
- Add Options.ApplyDefaults/ValidateRateLimitModes/TargetStrings helpers
- Extract ScanContext for runtime callback separation
- Delete MMutex, collectOrderedPocPaths, ~300 lines commented-out dead code
- Fix ReadComplieOptions typo → ReadCompileOptions
- Add 43 test cases across filter, context, poc, oob_coordinator, fingerprint
- Add ProgressRenderer + GracefulShutdown types (pending main.go integration)
- gofmt all files
