+++ pkg/runner/checker_brute_test.go
@@ -647,8 +647,8 @@ expression: oobCheckToken("dns", 3, "abc")
 			if err := yaml.Unmarshal([]byte(tt.yaml), p); err != nil {
 				t.Fatalf("unmarshal poc yaml: %v", err)
 			}
-			if got := checkerNeedsOOB(p); got != tt.want {
-				t.Fatalf("checkerNeedsOOB=%v, want %v", got, tt.want)
+			if got := config.PocUsesOOB(*p); got != tt.want {
+				t.Fatalf("config.PocUsesOOB=%v, want %v", got, tt.want)
 			}
 		})
 	}
